Nderrimi i emrave te tabelave defualt te identity dhe krijimi i DbContext

+++ backend/backend/Core/Entities/Log.cs
@@ -1,6 +1,6 @@
 ﻿namespace backend.Core.Entities
 {
-    public class Log : BaseEntity<int>
+    public class Log : BaseEntity<long>
     {
         public string? UserName { get; set; }
         public string Description { get; set; }
